Changes to `./src/chime_dash/app/` * Remove `components/container.py` and move addition of Sidebar to `components/__init__.py` * Remove `components/content.py` and move content to `pages/index.py` * Move `sidebar.py` from `components/` to `pages/` * Moved callbacks impacting data on index page into `pages/index.py` * Simplified all classes in `components/`
