+++ src/chime_dash/app/utils/__init__.py
@@ -7,3 +7,19 @@ Modules
 -------
 templates       utilities for localization templates
 """
+from . import callbacks
+from . import templates
+from itertools import repeat
+from urllib.parse import quote
+
+
+def build_csv_download(df):
+    return "data:text/csv;charset=utf-8,{csv}".format(csv=quote(df.to_csv(index=True, encoding='utf-8')))
+
+
+def toggle_hidden(boolean_input_value, elements_to_update):
+    result = []
+    for _ in repeat(None, elements_to_update):
+        # todo Fix once switch values make sense. Currently reported as 'None' for off and '[False]' for on
+        result.append(not boolean_input_value)
+    return result
