+++ src/chime_dash/app/utils/callbacks.py
@@ -1,4 +1,7 @@
 from dash import Dash
+from dash.exceptions import DuplicateCallbackOutput
+# todo Change to `from collections.abc import Iterable`. Using OrderedDict prevents multiple outputs to same DOM element
+# todo e.g., update children and toggle hidden
 from collections import OrderedDict
 from typing import Callable, List
 from functools import lru_cache
